+++ src/interfaces.ts
@@ -217,9 +217,24 @@ export interface CommentFlag {
     codes: DiagnosticCode[] | null;
 }
 
-export type CompilerPluginFactory = () => CompilerPlugin;
+export interface PluginFactoryOptions {
+    /**
+     * What version of brighterscript is activating this plugin? (Useful for picking different plugins or behavior based on the version of brighterscript)
+     */
+    version: string;
+}
+export type PluginFactory = (options?: PluginFactoryOptions) => Plugin;
+/**
+ * @deprecated use `PluginFactory` instead
+ */
+export type CompilerPluginFactory = PluginFactory;
 
-export interface CompilerPlugin {
+/**
+ * @deprecated use `Plugin` instead
+ */
+export type CompilerPlugin = Plugin;
+
+export interface Plugin {
     name: string;
     /**
      * Called before a new program is created
@@ -529,6 +544,10 @@ export interface AfterProgramValidateEvent extends BeforeProgramValidateEvent {
 export interface ProvideCompletionsEvent<TFile extends BscFile = BscFile> {
     program: Program;
     file: TFile;
+    /**
+     * The scopes this file is a member of. If the file is a member of no scopes, this will be an empty array.
+     * Plugins can use `event.program.globalScope` if the file is not a member of any scopes
+     */
     scopes: Scope[];
     position: Position;
     completions: CompletionItem[];