+++ src/util.ts
@@ -5,12 +5,12 @@ import { parse as parseJsonc, printParseErrorCode } from 'jsonc-parser';
 import * as path from 'path';
 import { rokuDeploy, DefaultFiles } from 'roku-deploy';
 import type { Diagnostic, Position, DiagnosticRelatedInformation } from 'vscode-languageserver';
-import { Location, Range } from 'vscode-languageserver';
+import { Range, Location } from 'vscode-languageserver';
 import { URI } from 'vscode-uri';
 import * as xml2js from 'xml2js';
 import type { BsConfig, FinalizedBsConfig } from './BsConfig';
 import { DiagnosticMessages } from './DiagnosticMessages';
-import type { CallableContainer, BsDiagnostic, FileReference, CallableContainerMap, CompilerPluginFactory, CompilerPlugin, ExpressionInfo, TranspileResult, MaybePromise, DisposableLike, ExtraSymbolData, GetTypeOptions, TypeChainProcessResult } from './interfaces';
+import type { CallableContainer, BsDiagnostic, FileReference, CallableContainerMap, Plugin, ExpressionInfo, TranspileResult, MaybePromise, DisposableLike, ExtraSymbolData, GetTypeOptions, TypeChainProcessResult, PluginFactory } from './interfaces';
 import { TypeChainEntry } from './interfaces';
 import { BooleanType } from './types/BooleanType';
 import { DoubleType } from './types/DoubleType';
@@ -55,6 +55,17 @@ import { InvalidType } from './types/InvalidType';
 export class Util {
     public clearConsole() {
         // process.stdout.write('\x1Bc');
+    }
+
+    /**
+     * Get the version of brighterscript
+     */
+    public getBrighterScriptVersion() {
+        try {
+            return fsExtra.readJsonSync(`${__dirname}/../package.json`).version;
+        } catch {
+            return undefined;
+        }
     }
 
     /**
@@ -788,46 +799,6 @@ export class Util {
     }
 
     /**
-     * Does the string appear to be a uri (i.e. does it start with `file:`)
-     */
-    public isUriLike(filePath: string) {
-        return filePath?.startsWith('file:');
-    }
-
-    /**
-     * Given a file path, convert it to a URI string
-     */
-    public pathToUri(filePath: string) {
-        if (!filePath) {
-            return filePath;
-        } else if (this.isUriLike(filePath)) {
-            return filePath;
-        } else {
-            return URI.file(filePath).toString();
-        }
-    }
-
-    /**
-     * Given a URI, convert that to a regular fs path
-     */
-    public uriToPath(uri: string) {
-        //if this doesn't look like a URI, then assume it's already a path
-        if (this.isUriLike(uri) === false) {
-            return uri;
-        }
-        let parsedPath = URI.parse(uri).fsPath;
-
-        //Uri annoyingly converts all drive letters to lower case...so this will bring back whatever case it came in as
-        let match = /\/\/\/([a-z]:)/i.exec(uri);
-        if (match) {
-            let originalDriveCasing = match[1];
-            parsedPath = originalDriveCasing + parsedPath.substring(2);
-        }
-        const normalizedPath = path.normalize(parsedPath);
-        return normalizedPath;
-    }
-
-    /**
      * Force the drive letter to lower case
      */
     public driveLetterToLower(fullPath: string) {
@@ -873,6 +844,46 @@ export class Util {
         }
         return true;
     }
+    /**
+     * Does the string appear to be a uri (i.e. does it start with `file:`)
+     */
+    public isUriLike(filePath: string) {
+        return filePath?.indexOf('file:') === 0;// eslint-disable-line @typescript-eslint/prefer-string-starts-ends-with
+    }
+
+    /**
+     * Given a file path, convert it to a URI string
+     */
+    public pathToUri(filePath: string) {
+        if (!filePath) {
+            return filePath;
+        } else if (this.isUriLike(filePath)) {
+            return filePath;
+        } else {
+            return URI.file(filePath).toString();
+        }
+    }
+
+    /**
+     * Given a URI, convert that to a regular fs path
+     */
+    public uriToPath(uri: string) {
+        //if this doesn't look like a URI, then assume it's already a path
+        if (this.isUriLike(uri) === false) {
+            return uri;
+        }
+        let parsedPath = URI.parse(uri).fsPath;
+
+        //Uri annoyingly converts all drive letters to lower case...so this will bring back whatever case it came in as
+        let match = /\/\/\/([a-z]:)/i.exec(uri);
+        if (match) {
+            let originalDriveCasing = match[1];
+            parsedPath = originalDriveCasing + parsedPath.substring(2);
+        }
+        const normalizedPath = path.normalize(parsedPath);
+        return normalizedPath;
+    }
+
 
     /**
      * Get the outDir from options, taking into account cwd and absolute outFile paths
@@ -1883,15 +1894,15 @@ export class Util {
     /**
      * Load and return the list of plugins
      */
-    public loadPlugins(cwd: string, pathOrModules: string[], onError?: (pathOrModule: string, err: Error) => void): CompilerPlugin[] {
+    public loadPlugins(cwd: string, pathOrModules: string[], onError?: (pathOrModule: string, err: Error) => void): Plugin[] {
         const logger = createLogger();
-        return pathOrModules.reduce<CompilerPlugin[]>((acc, pathOrModule) => {
+        return pathOrModules.reduce<Plugin[]>((acc, pathOrModule) => {
             if (typeof pathOrModule === 'string') {
                 try {
                     const loaded = requireRelative(pathOrModule, cwd);
-                    const theExport: CompilerPlugin | CompilerPluginFactory = loaded.default ? loaded.default : loaded;
+                    const theExport: Plugin | PluginFactory = loaded.default ? loaded.default : loaded;
 
-                    let plugin: CompilerPlugin | undefined;
+                    let plugin: Plugin | undefined;
 
                     // legacy plugins returned a plugin object. If we find that, then add a warning
                     if (typeof theExport === 'object') {
@@ -1900,7 +1911,9 @@ export class Util {
 
                         // the official plugin format is a factory function that returns a new instance of a plugin.
                     } else if (typeof theExport === 'function') {
-                        plugin = theExport();
+                        plugin = theExport({
+                            version: this.getBrighterScriptVersion()
+                        });
                     } else {
                         //this should never happen; somehow an invalid plugin has made it into here
                         throw new Error(`TILT: Encountered an invalid plugin: ${String(plugin)}`);
@@ -2069,7 +2082,7 @@ export class Util {
                 //filter out null relatedInformation items
             }).filter((x): x is DiagnosticRelatedInformation => Boolean(x)),
             code: diagnostic.code ? diagnostic.code : (diagnostic as BsDiagnostic).legacyCode,
-            source: 'brs'
+            source: diagnostic.source ?? 'brs'
         } as Diagnostic;
         if (diagnostic?.tags?.length > 0) {
             result.tags = diagnostic.tags;